DI Edge case: missing component
Removing the `@Service` annotation from our `SimpleGreetingService`
means it's not considered as a component anymore by Spring. Since we're
asking for it in our application, the startup phase fails with the
following message:

```
'fr.insalyon.tc.springbootintro.hello.GreetingService' that could
not be found.

Consider defining a bean of type
'fr.insalyon.tc.springbootintro.hello.GreetingService' in your configuration.
```