Implement expandShortCode
This method will do the exact opposite: for a given shortCode, we need
to be able to return the original `ShortURL` saved in the database (if
any).

We can reuse our repository here to find `ShortURL`s by their shortCode.
This method does not exist on our interface, but we can create it by
following the naming conventions of Spring Data.

Instead of dealing with `null` checks directly, we can leverage
`Optional` to check if the value was present in the database or throw an
exception if none was found.